src/main/java/com/arathy/transactionmanagement/service/TransactionService.java
package com.arathy.transactionmanagement.service;

import com.arathy.transactionmanagement.DAO.TransactionRepository;
import com.arathy.transactionmanagement.models.Transaction;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class TransactionService {
    TransactionRepository transactionRepository;

    @Autowired
    public TransactionService(TransactionRepository transactionRepository) {
        this.transactionRepository = transactionRepository;
    }

    public Transaction saveTransaction(Transaction transaction) {
        return transactionRepository.save(transaction);
    }
}
